Extract shared GraphGenerator base for DOT and Mermaid
DotGenerator and MermaidGenerator walked the resolved target graph
identically, only the per-node and per-edge text differed. Moved the
traversal and the shared helpers (label resolution, depfile parsing, ID
sanitization) into a GraphGenerator base, leaving each subclass with just
its header/footer and format-specific line formatters.

- Added pcons/generators/graph.py with the GraphGenerator base.
- Reduced dot.py and mermaid.py to format-specific hooks.
- Added DOT generation and verification to example 05, which previously
  exercised only Mermaid, so the DotGenerator path is now covered in CI.
